refactor : 알림 기능 전체적인 리팩터링
