+++ src/api/getMyNotofication.ts
@@ -1,9 +1,14 @@
 import axiosInstance from '@/lib/axiosInstance';
 import { QueryFunctionContext } from '@tanstack/react-query';
+import { NotificationDataType } from '@/types/notification';
 
-const getMyNotification = async ({ pageParam }: QueryFunctionContext): Promise<any> => {
+const getMyNotification = async ({ queryKey, pageParam }: QueryFunctionContext) => {
+  const [, size] = queryKey;
   const cursorParam = pageParam ? `&cursorId=${pageParam}` : '';
-  const response = await axiosInstance.get(`/my-notifications?size=4${cursorParam}`);
+  const dataSize = size ? `?size=${size}` : '';
+  const response = await axiosInstance.get<NotificationDataType>(
+    `/my-notifications${dataSize}${cursorParam}`,
+  );
   return response.data;
 };
