+++ src/components/header/NotificationDropdown.tsx
@@ -2,24 +2,10 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import getMyNotification from '@/api/getMyNotofication';
 import { useInView } from 'react-intersection-observer';
 import { useEffect } from 'react';
+import queryClient from '@/lib/queryClient';
+import { NotificationDataType } from '@/types/notification';
 import NotificationDropdownItem from './NotificationDropdownItem';
 import '../../styles/customScrollbar.css';
-
-interface Notifications {
-  id: number;
-  teamId: string;
-  userId: number;
-  content: string;
-  createdAt: string;
-  updatedAt: string;
-  deletedAt: string | null;
-}
-
-interface NotificationDataType {
-  totalCount: number;
-  notifications: Notifications[];
-  cursorId: number;
-}
 
 const NotificationDropdown = ({
   setDropdownIsOpen,
@@ -27,7 +13,7 @@ const NotificationDropdown = ({
   setDropdownIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }) => {
   const { data, fetchNextPage } = useInfiniteQuery<NotificationDataType>({
-    queryKey: ['notifications'],
+    queryKey: ['notifications', 10],
     queryFn: getMyNotification,
     initialPageParam: 0,
     getNextPageParam: (lastPage) => lastPage.cursorId,
@@ -36,6 +22,7 @@ const NotificationDropdown = ({
   const notifications = data?.pages.flatMap((page) => page.notifications) || [];
 
   const handleModalClose = () => {
+    queryClient.resetQueries({ queryKey: ['notifications', 10] });
     setDropdownIsOpen(false);
   };
   const { inView, ref } = useInView();
@@ -46,7 +33,7 @@ const NotificationDropdown = ({
   }, [inView, fetchNextPage]);
 
   return (
-    <div className="flex flex-col absolute top-12 right-12 z-20 w-[368px] rounded-md bg-green-10 shadow-md border-1 py-6 px-4 gap-3 h-[300px] overflow-y-auto custom-scrollbar">
+    <div className="flex flex-col absolute top-[63px] right-[-100px] md:right-[-30px] z-20 w-[368px] rounded-md bg-green-10 shadow-md border-1 py-6 px-4 gap-3 h-[300px] overflow-y-auto sm:w-screen sm:top-0 sm:right-0 sm:fixed sm:inset-0 sm:h-screen sm:rounded-none">
       {totalCount === 0 ? (
         <div>모든 알림을 확인했습니다!</div>
       ) : (
@@ -57,15 +44,19 @@ const NotificationDropdown = ({
               <img className="w-5 cursor-pointer" src="/assets/x_btn.svg" alt="Close Box Button" />
             </button>
           </div>
-          {notifications.map((item) => (
-            <NotificationDropdownItem
-              key={item.id}
-              content={item.content}
-              updatedAt={item.updatedAt}
-              notificationId={item.id}
-            />
-          ))}
-          <div ref={ref} className="w-[5px] h-[5px]" />
+          <ol className="overflow-y-auto custom-scrollbar flex flex-col w-[328px] rounded-md  gap-1 sm:w-full">
+            {notifications.map((item) => (
+              <NotificationDropdownItem
+                key={item.id}
+                content={item.content}
+                updatedAt={item.updatedAt}
+                notificationId={item.id}
+              />
+            ))}
+            <div ref={ref} className="w-[5px] h-[5px]">
+              &nbsp;
+            </div>
+          </ol>
         </>
       )}
     </div>
