+++ src/components/header/NotificationDropdownItem.tsx
@@ -1,43 +1,18 @@
 import queryClient from '@/lib/queryClient';
 import getDeleteNotification from '@/api/getDeleteNotification';
 import { useMutation } from '@tanstack/react-query';
+import formatUpdatedAt from '@/lib/utils/formatUpdatedAt';
+import { NotificationDropdownItemProps } from '@/types/notification';
 
-const formatUpdatedAt = (updatedAt: string) => {
-  const updatedDate = new Date(updatedAt);
-  const currentDate = new Date();
-  const diffInMs = currentDate.getTime() - updatedDate.getTime();
-  const diffInMin = Math.floor(diffInMs / (1000 * 60));
-  const diffInHours = Math.floor(diffInMin / 60);
-  const diffInDays = Math.floor(diffInHours / 24);
-  const diffInYears = Math.floor(diffInDays / 365);
-
-  if (diffInYears > 0) {
-    return `${diffInYears}년 전`;
-  }
-  if (diffInDays > 0) {
-    return `${diffInDays}일 전`;
-  }
-  if (diffInHours > 0) {
-    return `${diffInHours}시간 전`;
-  }
-  if (diffInMin > 0) {
-    return `${diffInMin}분 전`;
-  }
-  return '방금 전';
-};
 const NotificationDropdownItem = ({
   content,
   updatedAt,
   notificationId,
-}: {
-  content: string;
-  updatedAt: string;
-  notificationId: number;
-}) => {
+}: NotificationDropdownItemProps) => {
   const { mutate } = useMutation({
     mutationFn: getDeleteNotification,
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['notifications'] });
+      queryClient.invalidateQueries({ queryKey: ['notifications', 10] });
     },
   });
 
@@ -47,10 +22,11 @@ const NotificationDropdownItem = ({
 
   const reservationStatus = content.slice(content.length - 8, content.length - 6);
   return (
-    <div className="flex flex-col w-90 bg-white rounded-md p-2 gap-1">
+    <li className="w-full px-3 py-4 bg-white rounded-md sm:text-[20px]">
       <div className="flex justify-between">
         <img
           src={`/assets/circle_${reservationStatus === '승인' ? 'blue' : 'red'}.svg`}
+          className="sm:w-[10px] sm:h-[10px]"
           alt="Confirmed Chip"
         />
         <button type="button" onClick={onClickDeleteButton}>
@@ -73,8 +49,8 @@ const NotificationDropdownItem = ({
         </div>
         {content.split(reservationStatus)[1]}
       </div>
-      <div className="opacity-50 text-sm">{formatUpdatedAt(updatedAt)}</div>
-    </div>
+      <div className="opacity-50 text-sm mt-1 sm:text-[18px]">{formatUpdatedAt(updatedAt)}</div>
+    </li>
   );
 };
